converted the scripts file to ES modules, and added the ability to play/pause with the spacebar
