+++ index.html
@@ -17,5 +17,5 @@
                         <div class="progress"></div>
                 </div>
         </div>
-        <script type="module" src="scripts/scripts.js"></script>
+        <script type="module" src="scripts/main.js"></script>
 </body>
