+++ styles/styles.css
@@ -52,5 +52,5 @@ body {
   .cell img {
 	width: 60%;
 	left: 20%;
-}
+  }
 }
